refactor: adopt ResponseEntity for better API response

+++ src/main/java/com/oldiee/CatalogOnline/NotaController.java
@@ -1,5 +1,7 @@
 package com.oldiee.CatalogOnline;
 
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,20 +21,26 @@ public class NotaController {
     public List<Nota>getAllNotes(){
         return notaRepository.findAll();
     }
-    @GetMapping("/students/{studentId}")
+    @GetMapping("/student/{studentId}")
     public List<Nota>getNotesById(@PathVariable Long studentId){
         return notaRepository.findByStudentId(studentId);
     }
-    @PostMapping("/{studentId}/note")
-    public Nota addNota(@PathVariable Long studentId, @RequestBody Nota nota) {
-        return studentRepository.findById(studentId).map(student -> {
-            nota.setStudent(student);
-            return notaRepository.save(nota);
-        }).orElseThrow(() -> new RuntimeException("Studentul nu exista!"));
+    @PostMapping("/student/{studentId}")
+    public  ResponseEntity<Nota> addNota(@PathVariable Long studentId,@Valid @RequestBody Nota nota){
+        return  studentRepository.findById(studentId)
+                .map(student -> {
+                    nota.setStudent(student);
+                    Nota savedNota=notaRepository.save(nota);
+                    return  ResponseEntity.status(HttpStatus.CREATED).body(savedNota);
+                })
+                .orElse(ResponseEntity.notFound().build());
     }
     @DeleteMapping("/{id}")
     public ResponseEntity<Void>deleteNota(@PathVariable Long id){
+        if(notaRepository.existsById(id)){
         notaRepository.deleteById(id);
         return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
